refactor: 잘못된 구조 수정
테스트 메서드를 통과하기 위해 구조를 수정했습니다.

+++ src/main/java/lotto/service/InputService.java
@@ -19,7 +19,7 @@ public class InputService {
         }
     }
 
-    public List<BigInteger> getWinningNumbers() {
+    public List<Integer> getWinningNumbers() {
         InputView inputView = new InputView();
         Validator validator = new Validator();
